+++ src/main/java/Duke/Tasks/Commands.java
@@ -1,4 +1,4 @@
-package Tasks;
+package Duke.Tasks;
 
 public enum Commands {
     BYE,
