+++ src/main/java/Duke/Tasks/DeadlineTask.java
@@ -1,8 +1,6 @@
-package Tasks;
+package Duke.Tasks;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 
 public class DeadlineTask extends Task {
