+++ src/main/java/Duke/Tasks/TodoTask.java
@@ -1,4 +1,4 @@
-package Tasks;
+package Duke.Tasks;
 
 public class TodoTask extends Task {
