+++ text-ui-test/runtest.bat
@@ -7,7 +7,7 @@ REM delete output from previous run
 if exist .\text-ui-test\ACTUAL.TXT del .\text-ui-test\ACTUAL.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\bin .\src\main\java\*.java .\src\main\java\Tasks\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\bin .\src\main\java\*.java .\src\main\java\Duke.Tasks\*.java
 
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
@@ -16,7 +16,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Duke < .\text-ui-test\input.txt > .\text-ui-test\ACTUAL.TXT
+java -classpath ..\bin Duke.Duke < .\text-ui-test\input.txt > .\text-ui-test\ACTUAL.TXT
 
 REM compare the output to the expected output
 FC .\text-ui-test\ACTUAL.TXT .\text-ui-test\EXPECTED.TXT
